requirements.txt
asgiref==3.10.0
certifi==2025.10.5
charset-normalizer==3.4.4
dj-database-url==1.0.0
Django==4.2.25
django-filter==23.5
djangorestframework==3.16.1
djangorestframework-simplejwt==5.2.2
gunicorn==20.1.0
idna==3.11
psycopg==3.1.9
PyJWT==2.10.1
python-decouple==3.8
python-dotenv==1.0.0
requests==2.31.0
sqlparse==0.5.3
tzdata==2025.2
urllib3==2.5.0
whitenoise==6.5.0

